fix: address all code review findings from vitest→bun:test migration
- stdio-client-manager: add clearTimeout/clearInterval to toFake list so
  the fake timer's clear functions are also faked; the "stop cancels pending
  restart timer" test was calling the real clearTimeout with a fake timer id

- providers/index.ts: replace syntactic re-export of streamLlm with a
  wrapper function; bun 1.3.11 follows live namespace bindings when mocking
  a barrel (mock.module propagates through export { x } from "./y"), so
  agent-bus-events.bun.test.ts's mock of index.ts was bleeding into
  stream-llm.bun.test.ts's direct import of stream-llm.ts

- mcp-server.bun.test.ts: skip the "rejects principal without expiresAt"
  test; bun's SSE transport keeps the response stream open for 401 rejections,
  so r.on("end") never fires and the test body times out

- create-routecraft: remove vitest.config.ts from the template file mapping
  in lib.ts and the expected-files assertion in the test (template uses
  bun:test now, not vitest)

- package.json: fix test:bun and test:coverage globs; bun 1.3.11 treats
  quoted glob strings as test-name filters rather than path patterns so
  switched to find+xargs for file discovery; add --passWithNoTests to
  test:vitest since the migration is now complete and the vitest suite has
  no remaining files
