+++ package.json
@@ -22,10 +22,10 @@
     "lint": "eslint .",
     "format:write": "prettier --write . --ignore-unknown",
     "format": "prettier --check . --ignore-unknown",
-    "test": "bun run test:bun && bun run test:vitest",
-    "test:bun": "bun test bun.test",
-    "test:vitest": "vitest --run --exclude '**/integration.test.ts' --exclude '**/test/cross-runtime/**' --exclude '**/*.bun.test.ts' --exclude '**/*.bun.test.tsx'",
-    "test:coverage": "bun test bun.test --coverage --coverage-reporter=lcov --coverage-reporter=text --coverage-dir=coverage/bun && bun run test:vitest --coverage",
+    "test": "bun run test:bun; _bun=$?; bun run test:vitest; _vt=$?; test $_bun -eq 0 && test $_vt -eq 0",
+    "test:bun": "find packages/*/test examples/test \\( -name '*.bun.test.ts' -o -name '*.bun.test.tsx' \\) -not -path '*/cross-runtime/*' | sort | xargs bun test",
+    "test:vitest": "vitest --run --passWithNoTests --exclude '**/integration.test.ts' --exclude '**/test/cross-runtime/**' --exclude '**/*.bun.test.ts' --exclude '**/*.bun.test.tsx'",
+    "test:coverage": "find packages/*/test examples/test \\( -name '*.bun.test.ts' -o -name '*.bun.test.tsx' \\) -not -path '*/cross-runtime/*' | sort | xargs bun test --coverage --coverage-reporter=lcov --coverage-reporter=text --coverage-dir=coverage/bun && bun run test:vitest --coverage",
     "test:integration": "vitest --run 'integration.test.ts'",
     "test:cross-runtime": "vitest run --passWithNoTests test/cross-runtime/",
     "test:cross-runtime:node": "node node_modules/vitest/vitest.mjs run --passWithNoTests test/cross-runtime/",
