+++ packages/ai/src/llm/providers/index.ts
@@ -1,177 +1,38 @@
-import { logger as frameworkLogger } from "@routecraft/routecraft";
+import type { LlmResult } from "../types.ts";
+import {
+  buildExtras,
+  buildSdkParams,
+  collectToolCalls,
+  getStructuredOutput,
+  readReasoning,
+  toLlmUsage,
+  type CallLlmParams,
+  type ProviderExtras,
+} from "./llm-utils.ts";
+import { resolveLanguageModel } from "./resolve.ts";
+import { streamLlm as _streamLlm } from "./stream-llm.ts";
 import type { AgentDeltaListener } from "../../agent/events.ts";
-import { normalizeStreamDelta } from "../../agent/events.ts";
-import type {
-  LlmModelConfig,
-  LlmOptionsMerged,
-  LlmResult,
-  LlmToolCallSummary,
-  LlmUsage,
-} from "../types.ts";
 
-function throwProviderInstallError(pkg: string, provider: string): never {
-  throw new Error(
-    `The ${provider} LLM provider requires the "${pkg}" package. Install it with: bun add ${pkg}`,
-  );
-}
-
-function isModuleNotFoundFor(error: unknown, pkg: string): boolean {
-  if (!(error instanceof Error)) return false;
-  const msg = error.message ?? "";
-  return (
-    (msg.includes("ERR_MODULE_NOT_FOUND") ||
-      msg.includes("Cannot find module") ||
-      msg.includes("Cannot find package")) &&
-    msg.includes(pkg)
-  );
-}
-
-/** Provider-level defaults so users can register models with minimal config (e.g. { provider: "ollama" }). */
-const PROVIDER_DEFAULTS = {
-  ollama: {
-    baseURL: "http://localhost:11434/api",
-  },
-} as const;
-
-/** Pass through AI SDK usage so LlmUsage matches LanguageModelUsage (inputTokens/outputTokens). */
-function toLlmUsage(u: {
-  inputTokens?: number | undefined;
-  outputTokens?: number | undefined;
-  totalTokens?: number | undefined;
-}): LlmUsage {
-  return {
-    ...(u.inputTokens !== undefined && { inputTokens: u.inputTokens }),
-    ...(u.outputTokens !== undefined && { outputTokens: u.outputTokens }),
-    ...(u.totalTokens !== undefined && { totalTokens: u.totalTokens }),
-  };
-}
+export type { CallLlmParams } from "./llm-utils.ts";
 
 /**
- * Safely read structured output from generateText result. The AI SDK's result.output
- * is a getter that throws AI_NoOutputGeneratedError when the model didn't produce
- * valid structured output (e.g. empty, blocked, or unparseable). Catching here
- * allows returning without output so the adapter can try parsing result.text.
+ * Wrapper exported from the barrel so that `agent-bus-events.bun.test.ts`
+ * can mock this barrel without bun 1.3.11 propagating the mock through the
+ * syntactic re-export chain into `stream-llm.ts` itself.
+ * `stream-llm.bun.test.ts` imports directly from `stream-llm.ts` and must
+ * get the real implementation; a syntactic re-export here would cause bun to
+ * conflate the two modules in the process-global mock registry.
  */
-function getStructuredOutput(result: { output?: unknown }): unknown {
-  try {
-    if ("output" in result && result.output !== undefined) return result.output;
-  } catch {
-    // SDK getter threw (e.g. AI_NoOutputGeneratedError); leave output undefined.
-  }
-  return undefined;
-}
-
-/**
- * Runtime check that a provider-returned value has the minimal LanguageModel shape
- * required by the AI SDK generateText (doGenerate, doStream). Throws a descriptive
- * error if validation fails so we avoid silent type assertions.
- */
-function assertLanguageModelShape(
-  model: unknown,
-  providerName: string,
-  modelId: string,
-): void {
-  if (model === null || typeof model !== "object") {
-    throw new Error(
-      `[${providerName}] Invalid model: expected an object, got ${typeof model}. Model id: ${modelId}`,
-    );
-  }
-  const m = model as Record<string, unknown>;
-  if (typeof m["doGenerate"] !== "function") {
-    throw new Error(
-      `[${providerName}] Invalid model: missing or invalid doGenerate method. Model id: ${modelId}. ` +
-        "Ensure the provider returns an AI SDK-compatible language model.",
-    );
-  }
-  if (typeof m["doStream"] !== "function") {
-    throw new Error(
-      `[${providerName}] Invalid model: missing or invalid doStream method. Model id: ${modelId}. ` +
-        "Ensure the provider returns an AI SDK-compatible language model.",
-    );
-  }
-}
-
-export interface CallLlmParams {
-  config: LlmModelConfig;
-  /** Model id to use for the request (for OpenRouter may be config.modelId ?? registeredId). */
-  modelId: string;
-  options: Pick<
-    LlmOptionsMerged,
-    | "temperature"
-    | "maxTokens"
-    | "topP"
-    | "frequencyPenalty"
-    | "presencePenalty"
-  >;
-  system: string;
-  /**
-   * User-side conversation. When a string, it is sent as a single
-   * user prompt. When an array, it is forwarded directly to the SDK
-   * as the `prompt` argument (Vercel AI SDK accepts
-   * `string | Array<ModelMessage>` for `prompt`). The agent session
-   * uses the array form to feed back the prior assistant + tool
-   * messages plus a validator-corrective user message on a `validate`
-   * retry, so the model sees the full history rather than a fresh
-   * conversation.
-   */
-  user: string | unknown[];
-  /** Optional structured output spec (from toAiOutputSpec). Enables provider-level JSON schema. */
-  output?: unknown;
-  /**
-   * Optional Vercel AI SDK tool map. When supplied the SDK runs the
-   * tool-calling loop: presents tools to the model, dispatches calls,
-   * feeds results (and validation/guard errors) back, until the loop
-   * terminates (a final text response or `stopWhen` fires).
-   */
-  tools?: Record<string, unknown>;
-  /**
-   * Optional stop condition for the tool-calling loop. Required when
-   * `tools` is set; ignored otherwise. Built via `stepCountIs(n)` (or
-   * any other Vercel AI SDK stop predicate).
-   */
-  stopWhen?: unknown;
-  /**
-   * Optional abort signal forwarded into `generateText`. When the
-   * signal aborts mid-call, the SDK throws an AbortError and any
-   * in-flight tool handlers receive the same signal via their
-   * `FnHandlerContext.abortSignal`.
-   *
-   * Thread the route's signal here (`getExchangeRoute(exchange)?.signal`)
-   * so an in-flight agent dispatch is cancelled when the route or
-   * context shuts down.
-   */
-  abortSignal?: AbortSignal;
-}
-
-/**
- * Per-provider extras forwarded into `generateText` / `streamText`.
- * Centralised so each path builds the same shape and the typecast at
- * the call site stays narrow.
- *
- * @internal
- */
-interface ProviderExtras {
-  output?: unknown;
-  tools?: Record<string, unknown>;
-  stopWhen?: unknown;
-  abortSignal?: AbortSignal;
-}
-
-function buildExtras(params: CallLlmParams): ProviderExtras {
-  const out: ProviderExtras = {};
-  if (params.output !== undefined) out.output = params.output;
-  if (params.tools !== undefined && Object.keys(params.tools).length > 0) {
-    out.tools = params.tools;
-    if (params.stopWhen !== undefined) out.stopWhen = params.stopWhen;
-  }
-  if (params.abortSignal !== undefined) out.abortSignal = params.abortSignal;
-  return out;
+export async function streamLlm(
+  params: CallLlmParams & { onDelta: AgentDeltaListener },
+): Promise<LlmResult> {
+  return _streamLlm(params);
 }
 
 /**
  * Dispatch via Vercel AI SDK `generateText` and return a normalised
- * {@link LlmResult}. Resolves the language model for the configured
- * provider, then delegates to the shared {@link runGenerate} helper.
+ * `LlmResult`. Resolves the language model for the configured
+ * provider, then delegates to the shared `runGenerate` helper.
  */
 export async function callLlm(params: CallLlmParams): Promise<LlmResult> {
   const { config, modelId, options, system, user } = params;
@@ -179,216 +40,6 @@ export async function callLlm(params: CallLlmParams): Promise<LlmResult> {
   return runGenerate(model, options, system, user, buildExtras(params));
 }
 
-/**
- * Streaming counterpart to {@link callLlm}. Resolves the language model
- * for the configured provider, calls Vercel's `streamText`, forwards
- * each normalised token-level delta to `onDelta`, and finally returns
- * the consolidated {@link LlmResult} once the stream drains.
- *
- * Used by the agent destination when the user supplies
- * `AgentOptions.onDelta`. Exposed at the LLM-provider layer so the
- * provider plumbing (model resolution, option mapping, reasoning
- * extraction) stays in one place.
- */
-export async function streamLlm(
-  params: CallLlmParams & { onDelta: AgentDeltaListener },
-): Promise<LlmResult> {
-  const { config, modelId, options, system, user, onDelta } = params;
-  const model = await resolveLanguageModel(config, modelId);
-  return runStreamGenerate(
-    model,
-    options,
-    system,
-    user,
-    buildExtras(params),
-    onDelta,
-  );
-}
-
-/**
- * Resolve the AI SDK `LanguageModel` for a given provider config.
- * Single source of truth used by both the synchronous (`callLlm` →
- * `runGenerate`) and streaming (`streamLlm` → `runStreamGenerate`)
- * paths so provider setup is not duplicated per dispatch mode.
- *
- * @internal
- */
-async function resolveLanguageModel(
-  config: LlmModelConfig,
-  modelId: string,
-): Promise<unknown> {
-  switch (config.provider) {
-    case "openai":
-      return resolveOpenAI(config, modelId);
-    case "anthropic":
-      return resolveAnthropic(config, modelId);
-    case "gemini":
-      return resolveGemini(config, modelId);
-    case "openrouter":
-      return resolveOpenRouter(config, modelId);
-    case "ollama":
-      return resolveOllama(config, modelId);
-    default: {
-      const _: never = config;
-      throw new Error(
-        `LLM provider not implemented: ${(_ as LlmModelConfig).provider}`,
-      );
-    }
-  }
-}
-
-async function resolveOpenAI(
-  config: import("../types.ts").LlmModelConfigOpenAI,
-  modelId: string,
-): Promise<unknown> {
-  let createOpenAI: (s: {
-    apiKey: string;
-    baseURL?: string;
-  }) => (m: string) => unknown;
-  try {
-    const mod = await import("@ai-sdk/openai");
-    createOpenAI = mod.createOpenAI as typeof createOpenAI;
-  } catch (error) {
-    if (isModuleNotFoundFor(error, "@ai-sdk/openai")) {
-      throwProviderInstallError("@ai-sdk/openai", "OpenAI");
-    }
-    throw error;
-  }
-  const settings: { apiKey: string; baseURL?: string } = {
-    apiKey: config.apiKey,
-  };
-  if (config.baseURL !== undefined) settings.baseURL = config.baseURL;
-  const openai = createOpenAI(settings);
-  return openai(modelId);
-}
-
-async function resolveAnthropic(
-  config: import("../types.ts").LlmModelConfigAnthropic,
-  modelId: string,
-): Promise<unknown> {
-  let createAnthropic: (s: { apiKey: string }) => (m: string) => unknown;
-  try {
-    const mod = await import("@ai-sdk/anthropic");
-    createAnthropic = mod.createAnthropic as typeof createAnthropic;
-  } catch (error) {
-    if (isModuleNotFoundFor(error, "@ai-sdk/anthropic")) {
-      throwProviderInstallError("@ai-sdk/anthropic", "Anthropic");
-    }
-    throw error;
-  }
-  const anthropic = createAnthropic({ apiKey: config.apiKey });
-  return anthropic(modelId);
-}
-
-async function resolveGemini(
-  config: import("../types.ts").LlmModelConfigGemini,
-  modelId: string,
-): Promise<unknown> {
-  let createGoogleGenerativeAI: (s: {
-    apiKey: string;
-  }) => (m: string) => unknown;
-  try {
-    const mod = await import("@ai-sdk/google");
-    createGoogleGenerativeAI =
-      mod.createGoogleGenerativeAI as typeof createGoogleGenerativeAI;
-  } catch (error) {
-    if (isModuleNotFoundFor(error, "@ai-sdk/google")) {
-      throwProviderInstallError("@ai-sdk/google", "Gemini");
-    }
-    throw error;
-  }
-  const google = createGoogleGenerativeAI({ apiKey: config.apiKey });
-  return google(modelId);
-}
-
-async function resolveOpenRouter(
-  config: import("../types.ts").LlmModelConfigOpenRouter,
-  modelId: string,
-): Promise<unknown> {
-  let createOpenRouter: (s: { apiKey: string }) => {
-    chat: (id: string) => unknown;
-  };
-  try {
-    const mod = await import("@openrouter/ai-sdk-provider");
-    createOpenRouter = mod.createOpenRouter as typeof createOpenRouter;
-  } catch (error) {
-    if (isModuleNotFoundFor(error, "@openrouter/ai-sdk-provider")) {
-      throwProviderInstallError("@openrouter/ai-sdk-provider", "OpenRouter");
-    }
-    throw error;
-  }
-  const openrouter = createOpenRouter({ apiKey: config.apiKey });
-  const resolvedId = config.modelId ?? modelId;
-  const rawModel = openrouter.chat(resolvedId);
-  assertLanguageModelShape(rawModel, "OpenRouter", resolvedId);
-  return rawModel;
-}
-
-async function resolveOllama(
-  config: import("../types.ts").LlmModelConfigOllama,
-  modelId: string,
-): Promise<unknown> {
-  let createOllama: (s: { baseURL?: string }) => (name: string) => unknown;
-  try {
-    const mod = await import("ollama-ai-provider-v2");
-    createOllama = mod.createOllama as typeof createOllama;
-  } catch (error) {
-    if (isModuleNotFoundFor(error, "ollama-ai-provider-v2")) {
-      throwProviderInstallError("ollama-ai-provider-v2", "Ollama");
-    }
-    throw error;
-  }
-  const ollama = createOllama({
-    baseURL: config.baseURL ?? PROVIDER_DEFAULTS.ollama.baseURL,
-  });
-  const name = config.modelId ?? modelId;
-  const rawModel = ollama(name);
-  assertLanguageModelShape(rawModel, "Ollama", name);
-  return rawModel;
-}
-
-/**
- * Build the common `generateText` / `streamText` argument shape from
- * the merged options + system/user prompts + extras. Both the
- * synchronous and streaming paths feed this into their respective SDK
- * call so option mapping (temperature, max tokens, etc.) lives in one
- * place.
- *
- * @internal
- */
-function buildSdkParams(
-  model: unknown,
-  options: CallLlmParams["options"],
-  system: string,
-  user: string | unknown[],
-  extras: ProviderExtras,
-): Record<string, unknown> {
-  const params: Record<string, unknown> = {
-    model,
-    prompt: user,
-    temperature: options.temperature,
-  };
-  if (options.maxTokens !== undefined)
-    params["maxOutputTokens"] = options.maxTokens;
-  if (system) params["system"] = system;
-  if (options.topP !== undefined) params["topP"] = options.topP;
-  if (options.frequencyPenalty !== undefined) {
-    params["frequencyPenalty"] = options.frequencyPenalty;
-  }
-  if (options.presencePenalty !== undefined) {
-    params["presencePenalty"] = options.presencePenalty;
-  }
-  return { ...params, ...extras };
-}
-
-/**
- * Shared `generateText` invocation. Each provider helper builds the
- * model and delegates here so the genParams shape, extras handling
- * (output / tools / stopWhen), and result normalisation (text /
- * output / reasoning / usage) live in one place.
- *
- * @internal
- */
 async function runGenerate(
   model: unknown,
   options: CallLlmParams["options"],
@@ -407,7 +58,6 @@ async function runGenerate(
   if (parsed !== undefined) out.output = parsed;
   const reasoning = readReasoning(result);
   if (reasoning) out.reasoning = reasoning;
-  // generateText resolves finishReason synchronously on the result.
   const finishReason = (result as { finishReason?: unknown }).finishReason;
   if (typeof finishReason === "string") out.finishReason = finishReason;
   const toolCalls = collectToolCalls(result);
@@ -418,188 +68,4 @@ async function runGenerate(
     .response?.messages;
   if (Array.isArray(responseMessages)) out.responseMessages = responseMessages;
   return out;
-}
-
-/**
- * Shared `streamText` invocation. Iterates the SDK's `fullStream`,
- * forwards each token-level delta to `onDelta`, then awaits the
- * consolidated values (text, usage, reasoning, optional structured
- * output) once the stream drains.
- *
- * Listener errors are caught and logged; they do not abort the
- * dispatch. Stream errors propagate out of `await result.text`, so
- * the dispatch fails normally; coarse decision events (tool calls,
- * finish) are emitted on the context bus by the agent session, not
- * here.
- *
- * @internal
- */
-async function runStreamGenerate(
-  model: unknown,
-  options: CallLlmParams["options"],
-  system: string,
-  user: string | unknown[],
-  extras: ProviderExtras,
-  onDelta: AgentDeltaListener,
-): Promise<LlmResult> {
-  const { streamText } = await import("ai");
-  const params = buildSdkParams(model, options, system, user, extras);
-  const result = streamText(params as Parameters<typeof streamText>[0]);
-
-  for await (const part of result.fullStream) {
-    const delta = normalizeStreamDelta(part);
-    if (delta === null) continue;
-    try {
-      await onDelta(delta);
-    } catch (err) {
-      frameworkLogger.warn(
-        { err },
-        "agent.onDelta listener threw; ignoring and continuing stream",
-      );
-    }
-  }
-
-  // Drain consolidated values. `result.text` rejects with the same
-  // error that surfaced through fullStream, so failures propagate
-  // through this await.
-  const text = await result.text;
-  const out: LlmResult = { text: text ?? "", raw: result };
-  const usage = await safeAwait<{
-    inputTokens?: number | undefined;
-    outputTokens?: number | undefined;
-    totalTokens?: number | undefined;
-  }>(result.usage);
-  if (usage) out.usage = toLlmUsage(usage);
-  const reasoning = await safeAwait<string | undefined>(
-    (result as { reasoningText?: PromiseLike<string | undefined> })
-      .reasoningText,
-  );
-  if (typeof reasoning === "string" && reasoning.length > 0) {
-    out.reasoning = reasoning;
-  }
-  const structured = await safeAwait<unknown>(
-    (result as { output?: PromiseLike<unknown> }).output,
-  );
-  if (structured !== undefined) out.output = structured;
-  // streamText exposes finishReason as a Promise that resolves once
-  // the loop terminates; await it so callers (e.g. the agent
-  // session emitting `agent:finished`) see a normalised string
-  // instead of a Promise.
-  const finishReason = await safeAwait<string | undefined>(
-    (result as { finishReason?: PromiseLike<string | undefined> }).finishReason,
-  );
-  if (typeof finishReason === "string") out.finishReason = finishReason;
-  // streamText exposes `steps` as a Promise that resolves with the
-  // full step history once the loop terminates. Await + flatten into
-  // the same `LlmToolCallSummary[]` shape `runGenerate` produces, so
-  // callers see one normalised tool-call list across both paths.
-  const steps = await safeAwait<unknown>(
-    (result as { steps?: PromiseLike<unknown> }).steps,
-  );
-  const toolCalls = collectToolCalls({ steps });
-  if (toolCalls.length > 0) out.toolCalls = toolCalls;
-  if (Array.isArray(steps)) out.stepsCount = steps.length;
-  // streamText exposes the consolidated response (messages, body, ...)
-  // as a Promise that resolves once the stream drains. Await it so the
-  // session can append response.messages to a validate-retry prompt
-  // without re-walking the stream.
-  const response = await safeAwait<{ messages?: unknown }>(
-    (result as { response?: PromiseLike<{ messages?: unknown }> }).response,
-  );
-  if (response && Array.isArray(response.messages)) {
-    out.responseMessages = response.messages;
-  }
-  return out;
-}
-
-/**
- * Walk an SDK result's `steps` array (sync or post-await) and
- * produce a flat list of `LlmToolCallSummary` entries pairing each
- * tool call with its result or error. Tolerates missing fields and
- * legacy SDK shapes.
- *
- * @internal
- */
-function collectToolCalls(result: unknown): LlmToolCallSummary[] {
-  if (result === null || typeof result !== "object") return [];
-  const steps = (result as { steps?: unknown }).steps;
-  if (!Array.isArray(steps)) return [];
-  const out: LlmToolCallSummary[] = [];
-  for (const step of steps) {
-    if (step === null || typeof step !== "object") continue;
-    const calls = (step as { toolCalls?: unknown }).toolCalls;
-    const results = (step as { toolResults?: unknown }).toolResults;
-    if (!Array.isArray(calls)) continue;
-    for (const call of calls) {
-      if (call === null || typeof call !== "object") continue;
-      const c = call as Record<string, unknown>;
-      const toolCallId =
-        typeof c["toolCallId"] === "string" ? c["toolCallId"] : "";
-      const toolName = typeof c["toolName"] === "string" ? c["toolName"] : "";
-      const input = c["input"] ?? c["args"];
-      const summary: LlmToolCallSummary = {
-        toolCallId,
-        toolName,
-        input,
-      };
-      // Pair the matching result by toolCallId (preferred) or by
-      // identical position in the same step (legacy fallback).
-      if (Array.isArray(results)) {
-        const match = (results as Record<string, unknown>[]).find(
-          (r) => r["toolCallId"] === toolCallId,
-        );
-        if (match) {
-          if ("output" in match || "result" in match) {
-            summary.output = match["output"] ?? match["result"];
-          }
-          if ("error" in match && match["error"] !== undefined) {
-            summary.error = match["error"];
-          }
-        }
-      }
-      out.push(summary);
-    }
-  }
-  return out;
-}
-
-/**
- * Await a value that may be a `PromiseLike<T>` (the AI SDK uses
- * `PromiseLike` for stream consolidation accessors) and swallow
- * rejections, returning `undefined` instead. Used to read optional
- * accessors (usage, reasoning, structured output) where absence is
- * not an error.
- *
- * Rejections are logged at debug level so a real provider regression
- * (SDK shape change, transport error on the optional accessor itself)
- * leaves a breadcrumb without surfacing as a user-visible warning.
- */
-async function safeAwait<T>(
-  value: T | PromiseLike<T> | undefined,
-): Promise<T | undefined> {
-  if (value === undefined) return undefined;
-  try {
-    return await value;
-  } catch (err) {
-    frameworkLogger.debug(
-      { err },
-      "llm.streamLlm: optional accessor rejected; treating as absent",
-    );
-    return undefined;
-  }
-}
-
-/**
- * Defensive accessor for reasoning text. Vercel AI SDK exposes
- * `reasoningText` (concatenated string) when the provider returned
- * reasoning blocks (Anthropic extended thinking, OpenAI o1, etc.).
- * Some SDK versions or providers may omit it; treat absence as
- * "no reasoning available."
- */
-function readReasoning(result: unknown): string | undefined {
-  const r = result as { reasoningText?: unknown };
-  if (typeof r.reasoningText === "string" && r.reasoningText.length > 0) {
-    return r.reasoningText;
-  }
-  return undefined;
 }
