+++ packages/create-routecraft/templates/base/package.json
@@ -10,9 +10,8 @@
     "format": "prettier . --check --ignore-unknown",
     "format:write": "prettier . --write --ignore-unknown",
     "typecheck": "tsc --noEmit",
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "test:coverage": "vitest run --coverage"
+    "test": "bun test",
+    "test:coverage": "bun test --coverage"
   },
   "dependencies": {
     "@routecraft/routecraft": "ROUTECRAFT_VERSION"
@@ -23,12 +22,11 @@
     "@routecraft/testing": "ROUTECRAFT_VERSION",
     "@routecraft/eslint-plugin-routecraft": "ROUTECRAFT_VERSION",
     "@types/node": "^25.3.5",
-    "@vitest/coverage-v8": "^4.0.18",
+    "bun-types": "^1.1.0",
     "eslint": "^10.0.2",
     "pino-pretty": "^13.1.3",
     "prettier": "^3.8.1",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.56.1",
-    "vitest": "^4.0.18"
+    "typescript-eslint": "^8.56.1"
   }
 }
