+++ packages/create-routecraft/templates/examples/hello-world/capabilities/hello-world.bun.test.ts
@@ -15,7 +15,6 @@ describe("Hello World Routes", () => {
     if (t) {
       await t.stop();
     }
-    mock.restore();
   });
 
   /**
