+++ packages/routecraft/test/async-tap.bun.test.ts
@@ -9,7 +9,6 @@ describe("Async Tap Execution", () => {
     if (t) {
       await t.stop();
     }
-    mock.restore();
   });
 
   /**
